requirements.txt
guidance~=0.1.5
datasets~=2.16.1
omegaconf~=2.3.0
peft~=0.11.1
transformers~=4.41.2
#torch==2.1.1  # to build docker with required version of torch I had to comment this line
tqdm~=4.65.0
click~=8.1.7
PyYAML~=6.0.1
requests~=2.31.0
urllib3~=1.26.18
joblib~=1.2.0
configparser~=5.3.0
#setuptools==69.5.1 # to build docker with required version of setuptools I had to comment this line
mlflow~=2.9.2
bitsandbytes~=0.43.1
deepspeed~=0.12.4
hydra-core~=1.3.2
accelerate~=0.31.0
GitPython~=3.1.40
pytest~=7.4.3
xdict~=1.1.11
spacy~=3.7.4
numba==0.59.0
psutil~=5.9.0
sharedmem==0.3.8
trl~=0.9.4
evaluate~=0.4.0
scikit-learn~=1.4.0
pandas~=2.1.4
numpy~=1.26.3
networkx~=3.1
nltk~=3.8
spacy~=3.7
openai~=1.5
omegaconf~=2.3
jupyterlab==3.6.7
ipywidgets==8.0.4
build~=1.0
thefuzz~=0.22.1
